feat(vue): replace functions from Home component to WorkspaceContent
replaced functions created, generateDataCreated, convertQuotaForFrontend, added required imports in WorkspaceContent component
